Working on layout and update photo.
Updates to style to move the login down from the top of the page to a more usable place.
Changed the travis config to use node 4.2.x.
Deleted the unused users directory for a fresh start.
Page now loads error free and can get through the sign in and sign up.